Attempted to Fix some Test Cases
I got it to 17 Failing, 18 Passing (Had it with more working before I tried fixing some other issues, so maybe if data folder is deleted and you run it twice it'll be above 20 working). I've left some comments for areas I think might be causing some of the issues. One of my main issues is definitely coming from the tests attempting to run before the database is finished setting up after data folder is deleted. I don't know how to use wait-for-it.sh If it's already in a file somewhere, it doesn't seem to be working on my end. I will try to get back to it again tonight, and will definitely continue working on this tomorrow.